pypsa_tools/requirements.txt
# Testing Requirements for PyPSA MCP Server

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
pytest-xdist>=3.0.0  # For parallel test execution

# Mocking and fixtures
mock>=5.0.0
faker>=18.0.0
factory-boy>=3.2.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pylint>=2.17.0
isort>=5.12.0

# Performance testing
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0

# Test data generation
hypothesis>=6.70.0  # Property-based testing
numpy>=1.24.0
pandas>=2.0.0

# MCP testing utilities
fastmcp>=2.0.1  # Or whatever version you're using

# PyPSA and dependencies
pypsa>=0.25.0
highspy>=1.5.0  # Recommended solver
networkx>=3.0
cartopy>=0.21.0  # Optional, for plotting